requirements.txt
langchain-openai
langchain-core
langchain-community
streamlit
chainlit
ollama
python-dotenv
